+++ event-buffer/event-buffer-liquibase/src/main/rollback-scripts/stream_buffer-008-rollback.sql
@@ -1,0 +1,3 @@
+-- Roll back liquibase script '008-create-stream-error-table.changelog.xml'
+DROP TABLE stream_error;
+DELETE FROM databasechangelog WHERE id = 'stream_buffer-008';